Removed class LogicalDeviceType

+++ homegear-miscellaneous/src/MiscPeer.cpp
@@ -504,7 +504,7 @@ bool MiscPeer::load(BaseLib::Systems::ICentral* central)
 		_rpcDevice = GD::family->getRpcDevices()->find(_deviceType, _firmwareVersion, -1);
 		if(!_rpcDevice)
 		{
-			GD::out.printError("Error loading Miscellaneous peer " + std::to_string(_peerID) + ": Device type not found: 0x" + BaseLib::HelperFunctions::getHexString((uint32_t)_deviceType.type()) + " Firmware version: " + std::to_string(_firmwareVersion));
+			GD::out.printError("Error loading Miscellaneous peer " + std::to_string(_peerID) + ": Device type not found: 0x" + BaseLib::HelperFunctions::getHexString(_deviceType) + " Firmware version: " + std::to_string(_firmwareVersion));
 			return false;
 		}
 		initializeTypeString();
